requirements.txt
asgiref==3.11.0
boto3==1.42.15
botocore==1.42.15
crispy-tailwind==1.0.3
Django==5.0
django-crispy-forms==2.5
django-js-asset==3.1.2
django-storages==1.14.6
django-taggit==6.1.0
django-tinymce==5.0.0
jmespath==1.0.1
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.2
whitenoise==6.11.0
